pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "mono-slam"
version = "0.1.0"
description = "Monocular SLAM with ORB-SLAM3 and Zenoh pub/sub"
readme = "README.md"
authors = [
  { name = "Miguel Villa Floran", email = "miguel.villafloran@gmail.com" },
]
requires-python = ">=3.11"

dependencies = [
  "numpy",
  "opencv-contrib-python",
  "orbslam3-python",
  "eclipse-zenoh",
  "rerun-sdk>=0.30",
  "monopriors",
  "beartype>=0.22.9",
  "jaxtyping>=0.3.9",
  "einops>=0.8.2",
]

[tool.uv.sources]
monopriors = { git = "https://github.com/pablovela5620/monoprior.git" }
onnxruntime-gpu = { index = "jetson-ai-lab" }

[tool.uv]
override-dependencies = ["onnxruntime-gpu ; platform_machine != 'aarch64'"]

[[tool.uv.index]]
name = "jetson-ai-lab"
url = "https://pypi.jetson-ai-lab.io/sbsa/cu130"
explicit = true

[project.scripts]
camera-pub = "mono_slam.camera_pub:main"
slam-sub = "mono_slam.slam_sub:main"
